Fix instance-initializers folder
