+++ tests/unit/instance-initializers/segment-test.js
@@ -1,24 +1,36 @@
 import Application from '@ember/application';
-import { run } from '@ember/runloop';
+
+import config from 'dummy/config/environment';
 import { initialize } from 'dummy/instance-initializers/segment';
 import { module, test } from 'qunit';
+import Resolver from 'ember-resolver';
+import { run } from '@ember/runloop';
 
 module('Unit | Instance Initializer | segment', function (hooks) {
   hooks.beforeEach(function () {
-    run(() => {
-      this.application = Application.create();
-      this.appInstance = this.application.buildInstance();
+    this.TestApplication = class TestApplication extends Application {
+      modulePrefix = config.modulePrefix;
+      podModulePrefix = config.podModulePrefix;
+      Resolver = Resolver;
+    };
+    this.TestApplication.instanceInitializer({
+      name: 'initializer under test',
+      initialize,
     });
+    this.application = this.TestApplication.create({
+      autoboot: false,
+    });
+    this.instance = this.application.buildInstance();
   });
 
   hooks.afterEach(function () {
-    run(this.appInstance, 'destroy');
+    run(this.instance, 'destroy');
     run(this.application, 'destroy');
   });
 
-  // Replace this with your real tests.
-  test('it works', function (assert) {
-    initialize(this.appInstance);
+  // TODO: Replace this with your real tests.
+  test('it works', async function (assert) {
+    await this.instance.boot();
 
     // you would normally confirm the results of the initializer here
     assert.ok(true);
